// GYMFRONTEND/src/app/core/interfaces/pack/pack.ts
export interface Pack {
            
           id:string;
           nameplan: string;
           description: string;
           trialdays: number;
           features: string; 
           timedays: number;
           image: string;
           cost: number;
           code: string;
           status: string;



}